Unify terminology for module/pattern variables
In the parent PR we unified the `to_unpack` and `module_variable` types.  This left something to be desired, as the `module_variable` terminology remained in some places. This PR removes the `to_unpack` terminology in favor of calling it `module_variable`; this makes sense with the joint removal of `type_unpacks` and leads to parallel code to `pattern_variable`.